change back from embedded entity to inheritance
Changes from inheritance to embedded entity is at
fbe850e50638b2ab27699ba7b0fdec555ee38aec commit. The reason for that
changes is to ensure the position of column
`{created,updated,deleted}At` is after all column of `ReadList`.

This advantages has drawback that we can't use
`{created,updated,deleted}At` column for order and query in `find`
method.

For now the drawback of inheritance before about column position is
already fixed using migration. The SQL query in migrations file is
ensure the position of the table column.